requirements.txt
 Django==2.2.7
 psycopg2-binary
 requests
 daphne
 django-crispy-forms
 django-npm
 django-static-fontawesome
 django-bootstrap-modal-forms
 channels
 docker